+++ src/main/java/org/thoughtcrime/securesms/ConversationActivity.java
@@ -1135,6 +1135,7 @@ public class ConversationActivity extends PassphraseRequiredActionBarActivity
     recipient = new Recipient(this, dcChat);
     glideRequests = GlideApp.with(this);
 
+    searchMenu = null; // reset search on new intent
     setInputPanelVisibility(true);
     initializeContactRequest();
   }
@@ -1857,7 +1858,11 @@ public class ConversationActivity extends PassphraseRequiredActionBarActivity
       initializeSecurity(isSecureText, isDefaultSms);
       setInputPanelVisibility(false);
       initializeContactRequest();
-      invalidateOptionsMenu();
+      if (searchMenu == null) {
+        // invalidateOptionsMenu collapses the search bar,
+        // so only call it when user is not in search mode
+        invalidateOptionsMenu();
+      }
     } else if ((eventId == DcContext.DC_EVENT_INCOMING_MSG
             || eventId == DcContext.DC_EVENT_MSG_READ)
         && event.getData1Int() == chatId) {