+++ src/main/java/org/thoughtcrime/securesms/calls/CallActivity.java
@@ -1116,6 +1116,8 @@ public class CallActivity extends AppCompatActivity {
 
       proceedAfterPermissions();
     }
+
+    CallCoordinator.getInstance(this).ensureServiceStartedFromForeground();
   }
 
   @Override