+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Fix: Mic will stay active during a call when screen is off on Android 14+
+* Fix: allow to pick image without requiring storage permission
 
 ## v2.57.0
 2026-07