+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix: allow to pick image without requiring storage permission
 * Fix: avoid overflow with big numbers in badge counters
 * Fix: make video players truly fullscreen
+* Fix: avoid ending search mode when a new message arrive to the chat
 
 ## v2.57.0
 2026-07