+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Delta Chat Android Changelog
 
+## Unreleased
+
+* Fix: Mic will stay active during a call when screen is off on Android 14+
+* Fix: allow to pick image without requiring storage permission
+* Fix: avoid overflow with big numbers in badge counters
+* Fix: make video players truly fullscreen
+* Fix: avoid ending search mode when a new message arrive to the chat
+
 ## v2.57.0
 2026-07